refactor(mcp): rename identity modules onto the inbound/outbound axis
request_identity.py becomes inbound_identity.py (it extracts the inbound bearer a caller presents) and session_identity.py becomes outbound_identity.py (it resolves the outbound httpx.Auth each session binds). The names now state the axis the docstrings already framed, so the pair reads as counterparts. No behavior change; the auth facade re-exports the same contract symbols.
