+++ packages/mcp/src/pipefy_mcp/auth/__init__.py
@@ -1,16 +1,16 @@
-"""Identity for the HTTP transport: inbound bearer validation and session identity."""
+"""Identity for the HTTP transport: inbound bearer validation and outbound session identity."""
 
-from pipefy_mcp.auth.request_identity import require_request_bearer
+from pipefy_mcp.auth.inbound_identity import require_request_bearer
+from pipefy_mcp.auth.outbound_identity import (
+    AuthSource,
+    RequestScopedIdentity,
+    StartupIdentity,
+)
 from pipefy_mcp.auth.resource_server import (
     JwtTokenVerifier,
     ResourceServer,
     ResourceServerAuth,
     build_resource_server_auth,
-)
-from pipefy_mcp.auth.session_identity import (
-    AuthSource,
-    RequestScopedIdentity,
-    StartupIdentity,
 )
 
 __all__ = [
